Refs #23739: Pentagon video, update imports, split tools code
