+++ docs/rst/tutorials/vulcanai/vulcanai_tutorials.rst
@@ -9,9 +9,7 @@ executing complex tasks by decomposing them into manageable sub-tasks in any env
 VulcanAI is completely compatible with Vulcanexus, and works marvelously to control and even debug ROS 2
 robots and applications.
 
-A demo of the potential of VulcanAI can be found in the following video:
-
-`VulcanAI controlling TurtleSim <>`_.
+The following VulcanAI tutorials show how to use most of VulcanAI's features, starting from the basics concepts and gradually increasing in complexity until a complete application is built.
 
 .. toctree::
     :maxdepth: 1
@@ -19,3 +17,16 @@ A demo of the potential of VulcanAI can be found in the following video:
     vulcanai_installation/vulcanai_installation.rst
     vulcanai_tools/vulcanai_tools.rst
     turtlesim/vulcanai_turtlesim.rst
+
+Quick Demo
+==========
+
+A simple demo of the potential of VulcanAI can be found in the following video:
+
+.. raw:: html
+
+      <div class="video-center">
+        <video controls playsinline muted loop>
+          <source src="../../../_static/resources/tutorials/vulcanai/pentagon.mp4">
+        </video>
+      </div>
